pyproject.toml
[tool.poetry]
name = "convolution-demo"
version = "0.1.0"
description = "Various demo applets to demonstrate convolution filters and neural networks"
authors = ["lakewood999 <lakewood999@gmail.com>"]
license = "AGPL-3.0-or-lateter"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Flask = "^2.1.2"
numpy = "^1.22.3"
gunicorn = "^20.1.0"
opencv-python = "^4.5.5"
Pillow = "^9.1.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
